backend/petcode/requirements.txt
appnope==0.1.0
awesome-slugify==1.6.5
backcall==0.1.0
cached-property==1.5.1
certifi==2019.6.16
chardet==3.0.4
decorator==4.4.0
dj-database-url==0.5.0
Django==2.2.4
django-admin-jqueryui112==1.12.1b5
django-appconf==1.0.3
django-cors-headers==3.1.0
django-filter==2.2.0
django-heroku==0.3.1
django-imagekit==4.0.2
django-phone-field==1.7.1
django-starcross-gallery==1.0.6
django-url-filter==0.3.13
django-zip-code-api==1.4.2
djangorestframework==3.10.2
enum-compat==0.0.2
gunicorn==19.9.0
idna==2.8
ipdb==0.12.2
ipython==7.7.0
ipython-genutils==0.2.0
jedi==0.14.1
libmagic==1.0
Markdown==3.1.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pilkit==2.0
Pillow==5.0.0
prompt-toolkit==2.0.9
psycopg2==2.8.3
psycopg2-binary==2.8.3
ptyprocess==0.6.0
Pygments==2.4.2
pytz==2019.2
regex==2019.6.8
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
Unidecode==0.4.21
urllib3==1.25.3
wcwidth==0.1.7
whitenoise==4.1.3
